+++ package.json
@@ -17,6 +17,10 @@
     "qs": "^6.5.1",
     "whatwg-fetch": "^3.6.2"
   },
+  "peerDependencies": {
+    "@types/react": "^16 | ^17",
+    "react": "^16 | ^17"
+  },
   "files": [
     "dist"
   ],
@@ -32,10 +36,12 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@testing-library/react": "^12.1.2",
     "@types/jest": "^27.0.2",
     "@types/js-cookie": "^3.0.0",
     "@types/lodash": "^4.14.175",
     "@types/qs": "^6.9.7",
+    "@types/react": "^17",
     "concurrently": "^3.5.1",
     "esbuild": "^0.13.5",
     "esbuild-jest": "^0.5.0",
@@ -43,6 +49,8 @@
     "eslint-config-clutter": "^1.0.0-beta.3",
     "jest": "^27.2.5",
     "lodash": "^4.17.4",
+    "react": "^17",
+    "react-dom": "^17",
     "rimraf": "^2.5.4",
     "typescript": "^4.4.4"
   }
